RELEASE V21
Bug fix for Linux to support web apps that export to folder which path includes a space

By default a PharoJS test case runs Pharo object on the Pharo VM. See clas PjAppTestCase and traits PjTAppRunsInJS and PjTAppRunsInPharo

PjTAppFolderSharingSubclass. This trait is useful when defining a hierarchy of classes that rely on the same HTML and related files (CSS, third party JS, ...).. The JS is exported in the same folder as its superclass. This is typically useful for experimenting a variante of an existing class.

Renamed PjBrowserApp and PjFileBasedBrowserAppto to PjWebApplication and PjFileBasedWebApp

Notify user that export is DONE after clicking on the menu "PharoJS Playground"/<appClassName>/export