Fixed scoping problem for global declarations
There was a bug where functions were being overwritten by blank
declarations from PjTranspiler. All of the functions were declared in
the local scope, which takes precedence over the global scope. So when
we tried to make a subclass with pj_makeSubclass, this refers to the
empty function. If it tries to push to this.subclasses, it errors out
as this.subclasses is undefined.

It's fixed by removing the code in PjTranspiler that makes these 
declarations.

Some of the redundant declarations that were removed include:
function ZeroDivide(){};
function Margin(){};
function PjUndefinedObject(){};
function LoSubjectSet(){};