I made methods named by <jsGenerator: #xxx:> take a parameter, which will be a transpiler. This is at the moment just prepatory to putting a #pharoJsSelectorPrefix class-side method in an application so that you can choose the prefix on an app-by-app basis.
Also added a <jsLiteralGenerator> pragma that is useful for methods that want to do a complex calculation on the Pharo side (access a file, transpile something, etc.) but only care about the final result on the JS side. Whatever the method returns when run on Pharo is what the JS method will return. This works for anything that is a JS literal: numbers, strings, characters, booleans, arrays, disctionaries. This is somewhat similar to <jsGenerator:> except that <jsLiteralGenerator> returns something useful on both sides, whereas <jsGenerated:> is just for creating the JS method.

DIfferentiated between #isLiteralJavascriptValue and #isFreeJavascriptValue. Free means that it is either literal, or that it's a Class or a closed BlockClosure, or a disctionary or array for which all elements are free.

General refactoring.