Cleaned up an occasional shutdown failure (socket had already shut down, but registered an error because we were tying to tell it to shut down).
Made pharoJsSelectorPrefix be a class-side method on an app so you can choose the selector prefix, rather than it being global to the image.

Created a dummy PjMargin class because Point has a method asMargin and Margin tried to pull in all of Morphic.

Modified addClassToConvert: to make it easier to track down errant classes that pull in extra code (like all of Morphic!)